moved the farming functions to the abstraction - and a small typo fix :)
