+++ crates/subspace-farmer/src/lib.rs
@@ -19,6 +19,7 @@
 pub(crate) mod commands; // TODO: remove this again (temporarily inserted)
 pub(crate) mod commitments;
 pub(crate) mod common;
+pub(crate) mod farming;
 pub(crate) mod identity;
 pub(crate) mod object_mappings;
 pub(crate) mod plot;
